리팩토링: 중복 코드를 추상 클래스로 옮겨라
