+++ src/main/java/com/johngrib/objects/_10_call/NightlyDiscountPhone.java
@@ -9,13 +9,12 @@ import java.util.List;
 /**
  * 심야 할인 요금제.
  */
-public class NightlyDiscountPhone {
+public class NightlyDiscountPhone extends AbstractPhone {
   private static final int LATE_NIGHT_HOUR = 22;
 
   private Money nightlyAmount;
   private Money regularAmount;
   private Duration seconds;
-  private List<Call> calls = new ArrayList<>();
 
   public NightlyDiscountPhone(Money nightlyAmount, Money regularAmount, Duration seconds) {
     this.nightlyAmount = nightlyAmount;
@@ -23,20 +22,11 @@ public class NightlyDiscountPhone {
     this.seconds = seconds;
   }
 
-  public Money calculateFee() {
-    Money result = Money.ZERO;
-
-    for (Call call : calls) {
-      result = result.plus(calculateCallFee(call));
-    }
-    return result;
-  }
-
-  private Money calculateCallFee(Call call) {
+  @Override
+  protected Money calculateCallFee(Call call) {
     if (call.getFrom().getHour() >= LATE_NIGHT_HOUR) {
       return nightlyAmount.times(call.getDuration().getSeconds() / seconds.getSeconds());
-    } else {
-      return regularAmount.times(call.getDuration().getSeconds() / seconds.getSeconds());
     }
+    return regularAmount.times(call.getDuration().getSeconds() / seconds.getSeconds());
   }
 }
