+++ src/test/java/com/johngrib/objects/_10_call/PhoneTest.java
@@ -13,8 +13,8 @@ import static org.junit.jupiter.api.Assertions.*;
 @DisplayName("Phone 클래스")
 class PhoneTest {
   @Nested
-  @DisplayName("calculateFee 메소드")
-  class Describe_calculateFee {
+  @DisplayName("calculateCallFee 메소드")
+  class Describe_calculateCallFee {
 
     @Nested
     @DisplayName("10초에 5원씩 부과되는 요금제에 가입했고")
@@ -24,8 +24,8 @@ class PhoneTest {
       }
 
       @Nested
-      @DisplayName("1분짜리 통화가 두 번 있었다면")
-      class Context_with_2_calls {
+      @DisplayName("1분짜리 통화를 했다면")
+      class Context_with_1_min_call {
 
         Call getFirstCall() {
           return new Call(
@@ -34,24 +34,10 @@ class PhoneTest {
           );
         }
 
-        Call getSecondCall() {
-          return new Call(
-                  LocalDateTime.of(2018, 1, 2, 12, 10, 0),
-                  LocalDateTime.of(2018, 1, 2, 12, 11, 0)
-          );
-        }
-
-        Phone givenUsedPhone() {
-          final Phone phone = givenPhone();
-          phone.call(getFirstCall());
-          phone.call(getSecondCall());
-          return phone;
-        }
-
         @Test
-        @DisplayName("60원을 리턴한다")
+        @DisplayName("30원을 리턴한다")
         void it_returns_calculated_won() {
-          assertEquals(givenUsedPhone().calculateFee(), Money.wons(60.0));
+          assertEquals(givenPhone().calculateCallFee(getFirstCall()), Money.wons(30.0));
         }
       }
     }
